requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.12
Django==2.2.5
django-cors-headers==3.10.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
dnspython==2.2.1
idna==3.3
mongoengine==0.24.2
pymongo==4.1.1
pytz==2021.3
requests==2.27.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
urllib3==1.26.10
